#502 refactor interface, fix integration test for lead-acid
