+++ docs/source/models/submodels/interface/butler_volmer.rst
@@ -1,7 +1,5 @@
-Base Model
-==========
+Butler-Volmer
+=============
 
 .. autoclass:: pybamm.interface.butler_volmer.BaseModel
     :members:
-
-
