--- docs/source/models/submodels/interface/butler_volmer/butler_volmer_lead_acid.rst
@@ -1,6 +1,0 @@
-Lead Acid
-=========
-
-.. autoclass:: pybamm.interface.butler_volmer.LeadAcid
-    :members:
-
