--- docs/source/models/submodels/interface/butler_volmer/butler_volmer_lithium_ion.rst
@@ -1,6 +1,0 @@
-Lithium-ion
-============
-
-.. autoclass:: pybamm.interface.butler_volmer.LithiumIon
-    :members:
-
