--- docs/source/models/submodels/interface/butler_volmer/index.rst
@@ -1,10 +1,0 @@
-Butler-Volmer
-==============
-
-.. toctree::
-  :maxdepth: 1
-
-  base_butler_volmer
-  butler_volmer_lithium_ion
-  butler_volmer_lead_acid
-  surface_form/index
