+++ docs/source/models/submodels/interface/interface_lead_acid.rst
@@ -1,0 +1,5 @@
+Lead Acid
+=========
+
+.. autoclass:: pybamm.interface.BaseLeadAcid
+    :members:
