+++ docs/source/models/submodels/interface/interface_lithium_ion.rst
@@ -1,0 +1,5 @@
+Lithium-Ion
+===========
+
+.. autoclass:: pybamm.interface.BaseLithiumIon
+    :members:
