+++ docs/source/models/submodels/interface/inverse_butler_volmer.rst
@@ -1,6 +1,5 @@
-Base Model
-==========
+Inverse Butler-Volmer
+=====================
 
 .. autoclass:: pybamm.interface.inverse_butler_volmer.BaseModel
     :members:
-
