--- docs/source/models/submodels/interface/inverse_butler_volmer/index.rst
@@ -1,8 +1,0 @@
-Inverse Butler-Volmer
-=====================
-
-.. toctree::
-
-  base_inverse_butler_volmer
-  inverse_butler_volmer_lithium_ion
-  inverse_butler_volmer_lead_acid
