--- docs/source/models/submodels/interface/inverse_butler_volmer/inverse_butler_volmer_lead_acid.rst
@@ -1,7 +1,0 @@
-Lead Acid
-=========
-
-.. autoclass:: pybamm.interface.inverse_butler_volmer.LeadAcid
-    :members:
-
-
