--- docs/source/models/submodels/interface/inverse_butler_volmer/inverse_butler_volmer_lithium_ion.rst
@@ -1,5 +1,0 @@
-Lithium-Ion
-===========
-
-.. autoclass:: pybamm.interface.inverse_butler_volmer.LithiumIon
-    :members:
