+++ pybamm/models/submodels/interface/__init__.py
@@ -1,3 +1,5 @@
 from .base_interface import BaseInterface
 from . import butler_volmer
 from . import inverse_butler_volmer
+from . import lead_acid
+from . import lithium_ion
