--- pybamm/models/submodels/interface/butler_volmer/__init__.py
@@ -1,3 +1,0 @@
-from .base_bulter_volmer import BaseModel
-from .butler_volmer_lead_acid import LeadAcid
-from .butler_volmer_lithium_ion import LithiumIon
