--- pybamm/models/submodels/interface/inverse_butler_volmer/__init__.py
@@ -1,3 +1,0 @@
-from .base_inverse_butler_volmer import BaseModel
-from .inverse_butler_volmer_lead_acid import LeadAcid
-from .inverse_butler_volmer_lithium_ion import LithiumIon
