+++ pybamm/models/submodels/interface/lithium_ion.py
@@ -1,13 +1,13 @@
 #
-# Lithium ion inverse bulter-volmer class
+# Lithium-ion interface classes
 #
+from .base_interface import BaseInterface
+from . import butler_volmer, inverse_butler_volmer
 
-from .base_inverse_butler_volmer import BaseModel
 
-
-class LithiumIon(BaseModel):
+class BaseModel(BaseInterface):
     """
-    Lithium ion inverse Butler-Volmer class
+    Base lead-acid interface class
 
     Parameters
     ----------
@@ -24,7 +24,20 @@ class LithiumIon(BaseModel):
         super().__init__(param, domain)
 
     def _get_exchange_current_density(self, variables):
+        """
+        A private function to obtain the exchange current density for a lead acid
+        deposition reaction.
 
+        Parameters
+        ----------
+        variables: dict
+        `   The variables in the full model.
+
+        Returns
+        -------
+        j0 : :class: `pybamm.Symbol`
+            The exchange current density.
+        """
         c_s_surf = variables[self.domain + " particle surface concentration"]
         c_e = variables[self.domain + " electrolyte concentration"]
 
@@ -39,3 +52,13 @@ class LithiumIon(BaseModel):
         )
 
         return j0
+
+
+class ButlerVolmer(BaseModel, butler_volmer.BaseModel):
+    def __init__(self, param, domain):
+        super().__init__(param, domain)
+
+
+class InverseButlerVolmer(BaseModel, inverse_butler_volmer.BaseModel):
+    def __init__(self, param, domain):
+        super().__init__(param, domain)
