+++ tests/integration/test_old_models/test_old_submodels/test_old_interface.py
@@ -88,231 +88,6 @@ class TestOldHomogeneousReaction(unittest.TestCase):
             model.get_homogeneous_interfacial_current(None, "not a domain")
 
 
-class TestOldButlerVolmer(unittest.TestCase):
-    def setUp(self):
-        self.eta_r_n = pybamm.Variable("overpotential", ["negative electrode"])
-        self.eta_r_p = pybamm.Variable("overpotential", ["positive electrode"])
-        self.j0_n = pybamm.Variable("exchange-curr density", ["negative electrode"])
-        self.j0_p = pybamm.Variable("exchange-curr density", ["positive electrode"])
-
-    def tearDown(self):
-        del self.eta_r_n
-        del self.eta_r_p
-        del self.j0_n
-        del self.j0_p
-
-    def test_creation(self):
-        param = pybamm.standard_parameters_lithium_ion
-        model = pybamm.old_interface.OldInterfacialReaction(param)
-        j_n = model.get_butler_volmer(self.j0_n, self.eta_r_n)
-        j_p = model.get_butler_volmer(self.j0_p, self.eta_r_p)
-
-        # negative electrode Butler-Volmer is Multiplication
-        self.assertIsInstance(j_n, pybamm.Multiplication)
-        self.assertEqual(j_n.domain, ["negative electrode"])
-
-        # positive electrode Butler-Volmer is Multiplication
-        self.assertIsInstance(j_p, pybamm.Multiplication)
-        self.assertEqual(j_p.domain, ["positive electrode"])
-
-    def test_set_parameters(self):
-        param = pybamm.standard_parameters_lithium_ion
-        model = pybamm.old_interface.OldInterfacialReaction(param)
-        j_n = model.get_butler_volmer(self.j0_n, self.eta_r_n)
-        j_p = model.get_butler_volmer(self.j0_p, self.eta_r_p)
-
-        # Process parameters
-        parameter_values = model.default_parameter_values
-        j_n = parameter_values.process_symbol(j_n)
-        j_p = parameter_values.process_symbol(j_p)
-        [self.assertNotIsInstance(x, pybamm.Parameter) for x in j_n.pre_order()]
-        [self.assertNotIsInstance(x, pybamm.Parameter) for x in j_p.pre_order()]
-
-    def test_discretisation(self):
-        param = pybamm.standard_parameters_lithium_ion
-        model = pybamm.old_interface.OldInterfacialReaction(param)
-        j_n = model.get_butler_volmer(self.j0_n, self.eta_r_n)
-        j_p = model.get_butler_volmer(self.j0_p, self.eta_r_p)
-        j = pybamm.Concatenation(j_n, pybamm.Broadcast(0, ["separator"]), j_p)
-
-        # Process parameters and discretise
-        parameter_values = model.default_parameter_values
-        disc = get_discretisation_for_testing()
-        mesh = disc.mesh
-        disc.set_variable_slices([self.j0_n, self.eta_r_n, self.j0_p, self.eta_r_p])
-
-        j_n = disc.process_symbol(parameter_values.process_symbol(j_n))
-        j_p = disc.process_symbol(parameter_values.process_symbol(j_p))
-        j = disc.process_symbol(parameter_values.process_symbol(j))
-
-        # test butler-volmer in each electrode
-        submesh = np.concatenate(
-            [mesh["negative electrode"][0].nodes, mesh["positive electrode"][0].nodes]
-        )
-        y = np.concatenate([submesh ** 2, submesh ** 3])
-        self.assertEqual(
-            j_n.evaluate(None, y).shape, (mesh["negative electrode"][0].npts, 1)
-        )
-        self.assertEqual(
-            j_p.evaluate(None, y).shape, (mesh["positive electrode"][0].npts, 1)
-        )
-
-        # test concatenated butler-volmer
-        whole_cell = ["negative electrode", "separator", "positive electrode"]
-        whole_cell_mesh = disc.mesh.combine_submeshes(*whole_cell)
-        self.assertEqual(j.evaluate(None, y).shape, (whole_cell_mesh[0].npts, 1))
-
-    def test_diff_c_e_lead_acid(self):
-
-        # With intercalation
-        param = pybamm.standard_parameters_lead_acid
-        model = pybamm.old_interface_lead_acid.OldMainReaction(param)
-        parameter_values = model.default_parameter_values
-
-        def j_n(c_e):
-            eta_r_n = 1 - param.U_n(c_e)
-            return model.get_butler_volmer(c_e, eta_r_n, ["negative electrode"])
-
-        def j_p(c_e):
-            eta_r_p = 1 - param.U_p(c_e)
-            return model.get_butler_volmer(c_e, eta_r_p, ["positive electrode"])
-
-        c_e = pybamm.Scalar(0.5)
-        h = pybamm.Scalar(0.00001)
-
-        # Analytical
-        j_n_diff = parameter_values.process_symbol(j_n(c_e).diff(c_e))
-        j_p_diff = parameter_values.process_symbol(j_p(c_e).diff(c_e))
-
-        # Numerical
-        j_n_FD = parameter_values.process_symbol(
-            (j_n(c_e + h) - j_n(c_e - h)) / (2 * h)
-        )
-        self.assertAlmostEqual(j_n_diff.evaluate(), j_n_FD.evaluate())
-        j_p_FD = parameter_values.process_symbol(
-            (j_p(c_e + h) - j_p(c_e - h)) / (2 * h)
-        )
-        self.assertAlmostEqual(j_p_diff.evaluate(), j_p_FD.evaluate())
-
-    def test_diff_delta_phi_e_lead_acid(self):
-
-        # With intercalation
-        param = pybamm.standard_parameters_lead_acid
-        model = pybamm.old_interface_lead_acid.OldMainReaction(param)
-        parameter_values = model.default_parameter_values
-
-        def j_n(delta_phi):
-            return model.get_butler_volmer(1, delta_phi - 1, ["negative electrode"])
-
-        def j_p(delta_phi):
-            return model.get_butler_volmer(1, delta_phi - 1, ["positive electrode"])
-
-        delta_phi = pybamm.Scalar(0.5)
-        h = pybamm.Scalar(0.00001)
-
-        # Analytical
-        j_n_diff = parameter_values.process_symbol(j_n(delta_phi).diff(delta_phi))
-        j_p_diff = parameter_values.process_symbol(j_p(delta_phi).diff(delta_phi))
-
-        # Numerical
-        j_n_FD = parameter_values.process_symbol(
-            (j_n(delta_phi + h) - j_n(delta_phi - h)) / (2 * h)
-        )
-        self.assertAlmostEqual(j_n_diff.evaluate(), j_n_FD.evaluate())
-        j_p_FD = parameter_values.process_symbol(
-            (j_p(delta_phi + h) - j_p(delta_phi - h)) / (2 * h)
-        )
-        self.assertAlmostEqual(j_p_diff.evaluate(), j_p_FD.evaluate())
-
-    def test_failure(self):
-        model = pybamm.old_interface.OldInterfacialReaction(None)
-        with self.assertRaises(pybamm.DomainError):
-            model.get_butler_volmer(None, None, "not a domain")
-        with self.assertRaises(pybamm.DomainError):
-            model.get_inverse_butler_volmer(None, None, "not a domain")
-
-
-class TestOldExchangeCurrentDensity(unittest.TestCase):
-    def setUp(self):
-        self.c_e_n = pybamm.Variable("concentration", domain=["negative electrode"])
-        self.c_e_s = pybamm.Variable("concentration", domain=["separator"])
-        self.c_e_p = pybamm.Variable("concentration", domain=["positive electrode"])
-        self.c_e = pybamm.Concatenation(self.c_e_n, self.c_e_s, self.c_e_p)
-        self.c_s_n_surf = pybamm.surf(
-            pybamm.Variable("particle conc", domain=["negative particle"])
-        )
-        self.c_s_p_surf = pybamm.surf(
-            pybamm.Variable("particle conc", domain=["positive particle"])
-        )
-
-    def tearDown(self):
-        del self.c_e_n
-        del self.c_e_s
-        del self.c_e_p
-        del self.c_e
-        del self.c_s_n_surf
-        del self.c_s_p_surf
-
-    def test_creation_lithium_ion(self):
-        param = pybamm.standard_parameters_lithium_ion
-        model = pybamm.old_interface.OldLithiumIonReaction(param)
-        j0_n = model.get_exchange_current_densities(self.c_e_n, self.c_s_n_surf)
-        j0_p = model.get_exchange_current_densities(self.c_e_p, self.c_s_p_surf)
-        self.assertEqual(j0_n.domain, ["negative electrode"])
-        self.assertEqual(j0_p.domain, ["positive electrode"])
-
-    def test_set_parameters_lithium_ion(self):
-        param = pybamm.standard_parameters_lithium_ion
-        model = pybamm.old_interface.OldLithiumIonReaction(param)
-        j0_n = model.get_exchange_current_densities(self.c_e_n, self.c_s_n_surf)
-        j0_p = model.get_exchange_current_densities(self.c_e_p, self.c_s_p_surf)
-        # Process parameters
-        parameter_values = model.default_parameter_values
-        j0_n = parameter_values.process_symbol(j0_n)
-        j0_p = parameter_values.process_symbol(j0_p)
-        # Test
-        [self.assertNotIsInstance(x, pybamm.Parameter) for x in j0_n.pre_order()]
-        [self.assertNotIsInstance(x, pybamm.Parameter) for x in j0_p.pre_order()]
-
-    def test_discretisation_lithium_ion(self):
-        param = pybamm.standard_parameters_lithium_ion
-        model = pybamm.old_interface.OldLithiumIonReaction(param)
-        j0_n = model.get_exchange_current_densities(self.c_e_n, self.c_s_n_surf)
-        j0_p = model.get_exchange_current_densities(self.c_e_p, self.c_s_p_surf)
-        # Process parameters and discretise
-        parameter_values = model.default_parameter_values
-        disc = get_discretisation_for_testing()
-        mesh = disc.mesh
-        disc.set_variable_slices(
-            [self.c_e, self.c_s_n_surf.orphans[0], self.c_s_p_surf.orphans[0]]
-        )
-        j0_n = disc.process_symbol(parameter_values.process_symbol(j0_n))
-        j0_p = disc.process_symbol(parameter_values.process_symbol(j0_p))
-
-        # Test
-        whole_cell = ["negative electrode", "separator", "positive electrode"]
-        submesh = mesh.combine_submeshes(*whole_cell)
-        y = np.concatenate(
-            [
-                submesh[0].nodes ** 2,
-                mesh["negative particle"][0].nodes,
-                mesh["positive particle"][0].nodes,
-            ]
-        )
-        # should evaluate to vectors with the right shape
-        self.assertEqual(
-            j0_n.evaluate(y=y).shape, (mesh["negative electrode"][0].npts, 1)
-        )
-        self.assertEqual(
-            j0_p.evaluate(y=y).shape, (mesh["positive electrode"][0].npts, 1)
-        )
-
-    def test_failure(self):
-        model = pybamm.old_interface.OldLithiumIonReaction(None)
-        with self.assertRaises(pybamm.DomainError):
-            model.get_exchange_current_densities(None, None, "not a domain")
-
-
 if __name__ == "__main__":
     print("Add -v for more debug output")
     import sys
